% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Run application}
\usage{
launch_app()
}
\value{
An object that represents the app. Printing the object or passing it to runApp() will run the app.
}
\description{
Run application
}
